add article and test

+++ webook/internal/web/types.go
@@ -3,7 +3,7 @@ package web
 import "github.com/gin-gonic/gin"
 
 type Result struct {
-	Code int    `json:"code"`
+	Code int64  `json:"code"`
 	Msg  string `json:"msg"`
 	Data any    `json:"data"`
 }
